requirement.txt
azure-storage-blob==12.14.1
fastapi==0.116.1
numpy==2.3.2
pandas==2.3.2
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
uvicorn
